[refactor] modify to fit MVC framework
