+++ src/main/java/com/oneswap/repository/NetworkRepository.java
@@ -1,0 +1,12 @@
+package com.oneswap.repository;
+
+import org.springframework.stereotype.Repository;
+
+import java.math.BigInteger;
+
+@Repository
+public interface NetworkRepository {
+
+    void saveGasFee(BigInteger gasPrice);
+
+}
